+++ src/app/tezos/tezos-wallet/tezos-wallet-detail/tezos-wallet-detail.reducer.ts
@@ -24,12 +24,12 @@ export function reducer(state = initialState, action) {
         }
 
         case 'TEZOS_NODE_PRICE_UPDATE_SUCCESS': {
+
             return {
                 ...state,
-               // price: action.payload.XTZ.USD,
-               price: 0.4696898
-                // price: action.payload.data.quote.XTZ.USD.price,
+               price: action.payload.XTZ.USD,
             }
+            
         }
 
         case 'TEZOS_WALLET_DETAIL_NODE_DETAIL_SUCCESS': {