+++ src/app/tezos/tezos-wallet/tezos-wallet-list/tezos-wallet-list.effects.ts
@@ -3,7 +3,7 @@ import { Injectable, NgZone } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Observable, of } from 'rxjs';
-import { map, withLatestFrom, flatMap, catchError, filter, tap } from 'rxjs/operators';
+import { map, withLatestFrom, flatMap, switchMap, catchError, filter, tap, takeLast, first } from 'rxjs/operators';
 import { ofRoute, enterZone } from './../../../shared/utils/rxjs/operators';
 
 import { initializeWallet, getWallet } from 'tezos-wallet'
@@ -34,18 +34,32 @@ export class TezosWalletListEffects {
     @Effect()
     TezosWalletListLoad$ = this.actions$.pipe(
         ofType('TEZOS_WALLET_LIST_LOAD'),
+
+        tap(action => console.log('[TEZOS_WALLET_LIST_LOAD][action]', action)),
+
         // get state from store
         withLatestFrom(this.store, (action, state: any) => state),
 
+        tap(state => console.log('[TEZOS_WALLET_LIST_LOAD][state]', state)),
+
         // get data from firebase 
         // TODO: move to custom rxjs operator
-        flatMap(state =>
+        switchMap(state =>
+
             of([]).pipe(
+                tap(address => console.log('[TEZOS_WALLET_LIST_LOAD][start] address', address)),
+
                 // get data from firebase 
-                flatMap(() => this.db.collection('tezos_' + state.tezos.tezosNode.api.name + '_wallet',
+                // TODO: update data only after has final state this.db.collection
+                // it will reduce number of request to BE
+                switchMap(() => this.db.collection('tezos_' + state.tezos.tezosNode.api.name + '_wallet',
                     query => query.where('uid', '==', null)
-                        .orderBy('name', 'asc')
-                ).valueChanges()),
+                        .orderBy('name', 'asc'))
+                    .valueChanges()
+                ),
+
+                tap(address => console.log('[TEZOS_WALLET_LIST_LOAD] address', address)),
+
                 // show only valid trezor addresses or dektop with private key 
                 map(addresses => addresses
                     .filter((address: any) =>
@@ -75,8 +89,12 @@ export class TezosWalletListEffects {
     TezosWalletListBalanceUpdate$ = this.actions$.pipe(
         ofType('TEZOS_WALLET_LIST_LOAD_SUCCESS'),
 
+        tap(action => console.log('[TEZOS_WALLET_LIST_LOAD_SUCCESS][action]', action)),
+
         // get state from store
         withLatestFrom(this.store, (action, state: any) => state),
+
+        tap(state => console.log('[TEZOS_WALLET_LIST_LOAD_SUCCESS][state]', state)),
 
         // get all accounts address
         flatMap((state: any) => state.tezos.tezosWalletList.ids