+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - allow ctr-c ctrl-v 
 - fixed issue with XTZ/USD price update
 - fixed issue with external links 
+