+++ CHANGELOG.md
@@ -9,4 +9,6 @@
 - allow ctr-c ctrl-v 
 - fixed issue with XTZ/USD price update
 - fixed issue with external links 
+- added max allowed amount validation for origination
+